+++ frontend/src/components/tests/CourseSearch.spec.js
@@ -3,16 +3,17 @@ import { describe, it, expect } from 'vitest'
 import vuetify from '@/plugins/vuetify'
 import pinia from '@/store'
 import { mount } from '@vue/test-utils'
-import App from '@/App.vue'
+import CourseSearch from '../CourseSearch.vue'
 
 import ResizeObserver from 'resize-observer-polyfill'
 global.ResizeObserver = ResizeObserver
 
-describe('some test', () => {
-  it('initializes message_sent properly', () => {
-    const wrapper = mount(App, {
+describe('CourseSearch', () => {
+  it('renders properly', () => {
+    const wrapper = mount(CourseSearch, {
       global: { plugins: [vuetify, pinia] }
     })
-    expect(wrapper.text()).toContain('PlanDoskonały')
+    expect(wrapper.text()).toContain('dodaj przedmiot')
   })
 })
+
