--- frontend/src/store/app.js
@@ -1,8 +1,0 @@
-// Utilities
-import { defineStore } from 'pinia'
-
-export const useAppStore = defineStore('app', {
-  state: () => ({
-    //
-  })
-})
